Move test classes to namespaces
