+++ tests/Functional/ConfigurationTest.php
@@ -2,35 +2,16 @@
 
 declare(strict_types=1);
 
-namespace Basis\Nats\Tests;
+namespace Tests\Functional;
 
-use Basis\Nats\Configuration;
+use Tests\FunctionalTestCase;
 
-class ConfigurationTest extends Test
+class ConfigurationTest extends FunctionalTestCase
 {
     public function testClientConfigurationOverride()
     {
         $this->assertSame($this->getConfiguration()->host, getenv('NATS_HOST'));
         $this->assertEquals($this->getConfiguration()->port, getenv('NATS_PORT'));
-    }
-
-    public function testClientConfigurationToken()
-    {
-        $connection = new Configuration(['token' => 'zzz']);
-        $this->assertArrayHasKey('auth_token', $connection->getOptions());
-    }
-
-    public function testClientConfigurationJwt()
-    {
-        $connection = new Configuration(['jwt' => random_bytes(16)]);
-        $this->assertArrayHasKey('jwt', $connection->getOptions());
-    }
-
-    public function testClientConfigurationBasicAuth()
-    {
-        $connection = new Configuration(['user' => 'nekufa', 'pass' => 't0p53cr3t']);
-        $this->assertArrayHasKey('user', $connection->getOptions());
-        $this->assertArrayHasKey('pass', $connection->getOptions());
     }
 
     public function testStreamConfgurationInvalidStorageBackend()
